minor grammatical and style changes to preface.asc
Comma after "It is much more" to allow the sentence to flow better. "well-known" instead of "well known". Comma between before the "and" following "Coffeescript" to break up the sentence. Oxford comma in "...open...[files like computer programs], modify them, and save them again."

Oxford comma in "...customize, extend, and modify."

The "like an elementary student" seems just a bit awkward and unnecessary, especially since the following "...all the way to..." statement doesn't make another comparison (e.g. "like a pro" or something to that effect). The addition of a comma also makes the sentence easier to read ("...we'll cover everything from the basics..., all the way to...")